requirements.txt
streamlit>=1.28.0
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-chroma==0.1.4
langchain-groq==0.2.0
langchain-huggingface==0.1.2
chromadb==0.5.3
sentence-transformers>=2.6.0
pypdf==3.17.0
python-dotenv==1.0.0